requirements.txt
streamlit
tensorflow
numpy
opencv-python
Pillow
opencv-python-headless
